Adds support for having repo as a submodule
We need to know where the repo is located as a submodule.  This is
indicated using the variable INCLUDE_MAKEFILES.  This variable should
contain the relative path to the submodule root.  The makefiles in the
repo can then also rely on INCLUDE_MAKEFILES to avoid downloading their
dependencies instead of using those in the repo.

The same applies for using MIUN TeX in the same manner.  If the variable
INCLUDE_MIUNTEX is set, then it must contain the relative path to the
root of the miuntex submodule.  If set, then the makefiles will create
symbolic links to these files instead of downloading new copies.  Note
that the dependencies must be set for this to work correctly.