Updates biblatex-lncs
Check for the existence of a submodule biblatex-lncs, which is the real
repo name.  If that exists, use it.  Otherwise, clone the newest version
from GitHub.

Use files by symbolic links.  Then any modifications will be detected in
the actual repo.

After looking into the fork history, I found newer versions of
biblatex-lncs.