Improves NOWEB handling
Now we use the `-delay` option for NOWEB only for .nw files.  This is
assumed to be a bigger file which uses its own document class etc. and
includes other files.

The files to be included are more specific files such as .cpp.nw, which
focus on a smaller part of a program.