Temporarily disables miun.*.mk files
Since the main include files have changed considerably, the miun.port.mk
will not work as backwards compatible.

The way this should be fixed is by rewriting the miun.*.mk files using
the new include files, i.e. basically to map old variable names to the
new ones to achieve the old behaviour.

This is further discussed in #23.