Removes unneeded cpif for py and mk files
Since Python and make(1) are non-compilable languages, we do not need
the cpif(1) utility. The cpif(1) utility is needed only for compilable
languages, e.g. C++, to prevent a lot of recompilation. For example,
a header file might not have been updated, but will be rebuilt by
noweb(1) if the .cpp file was updates, and thus cause a lot of unneeded
recompilation.