Updates structure for tests.mk (#25)
We now introduce `tests.mk` in the intro chapter. However, the tests
will be covered in each chapter (i.e. in the NOWEB file of each
include).

There is a `tests.nw` file which sets up the structure. This file is
covered in the introduction. However, each file, e.g. `tex.mk.nw`, will
also have a code block `<<tests.mk>>` containing all test code. Then
`tests.mk` can be constructed from `tests.nw tex.mk.nw` and so on.

The rationale is that the tests should be with the code. Otherwise they
risk getting out of sync.